add: timelines, homepage

+++ crates/core/src/database/auth.rs
@@ -139,7 +139,7 @@ impl DataManager {
             Err(e) => return Err(Error::DatabaseConnection(e.to_string())),
         };
 
-        let res = execute!(&conn, "DELETE FROM users WHERE id = $1", &[&(id as i64)]);
+        let res = execute!(&conn, "DELETE FROM users WHERE id = $1", &[&(id as isize)]);
 
         if let Err(e) = res {
             return Err(Error::DatabaseError(e.to_string()));
