+++ crates/core/src/database/userblocks.rs
@@ -39,7 +39,7 @@ impl DataManager {
         let res = query_row!(
             &conn,
             "SELECT * FROM userblocks WHERE initiator = $1 AND receiver = $2",
-            &[&(initiator as i64), &(receiver as i64)],
+            &[&(initiator as isize), &(receiver as isize)],
             |x| { Ok(Self::get_userblock_from_row(x)) }
         );
 
@@ -64,7 +64,7 @@ impl DataManager {
         let res = query_row!(
             &conn,
             "SELECT * FROM userblocks WHERE receiver = $1 AND initiator = $2",
-            &[&(receiver as i64), &(initiator as i64)],
+            &[&(receiver as isize), &(initiator as isize)],
             |x| { Ok(Self::get_userblock_from_row(x)) }
         );
